requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==0.19.0
SQLAlchemy==2.0.0
typing-extensions==4.4.0
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.12.0
